scalping_v2/requirements.txt
# ADX Trading Strategy v2.0 - Python Dependencies
# Created: 2025-10-15

# Core libraries
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
TA-Lib>=0.4.28
pandas-ta>=0.3.14b0

# Database
mysql-connector-python>=8.0.33
PyMySQL>=1.1.0
SQLAlchemy>=2.0.0

# API & Networking
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Data validation
pydantic>=2.0.0

# Logging & Monitoring
colorlog>=6.7.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
